requirements.txt
boto==2.49.0
boto3==1.9.18
botocore==1.12.18
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
oauthlib==2.1.0
PySocks==1.6.8
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
tweepy==3.6.0
urllib3==1.23
